// frontend/src/types/Bowler.ts
export type Bowler = {
  team: any;
  bowlerId: number;
  bowlerLastName: string;
  bowlerFirstName: string;
  bowlerMiddleInit: string;
  bowlerAddress: string;
  bowlerCity: string;
  bowlerState: string;
  bowlerZip: string;
  bowlerPhoneNumber: string;
  teamId: number;
  teamName: string;
};

// export type Bowler = {
//   bowlerID: number;
//   bowlerLastName: string;
//   bowlerFirstName: string;
//   bowlerMiddleInit?: string;
//   bowlerAddress: string;
//   bowlerCity: string;
//   bowlerState: string;
//   bowlerZip: string;
//   bowlerPhoneNumber: string;
//   teamID: number;
//   teamName: string;
// };